+++ host-configs/docker/docker-linux-ubuntu16.04-ivybridge-gcc@8.1.0.cmake
@@ -32,9 +32,9 @@ set(BLT_CXX_STD "c++14" CACHE PATH "")
 # Root directory for generated TPLs
 set(TPL_ROOT "/home/axom/axom_tpls/gcc-8.1.0" CACHE PATH "")
 
-set(CONDUIT_DIR "${TPL_ROOT}/conduit-0.5.1" CACHE PATH "")
+set(CONDUIT_DIR "${TPL_ROOT}/conduit-0.6.0" CACHE PATH "")
 
-# MFEM not built
+set(MFEM_DIR "${TPL_ROOT}/mfem-4.1.0" CACHE PATH "")
 
 set(HDF5_DIR "${TPL_ROOT}/hdf5-1.8.21" CACHE PATH "")