+++ host-configs/docker/docker-linux-ubuntu18.04-x86_64-clang@10.0.0.cmake
@@ -30,9 +30,9 @@ set(BLT_CXX_STD "c++14" CACHE PATH "")
 # Root directory for generated TPLs
 set(TPL_ROOT "/home/axom/axom_tpls/clang-10.0.0" CACHE PATH "")
 
-set(CONDUIT_DIR "${TPL_ROOT}/conduit-0.5.1" CACHE PATH "")
+set(CONDUIT_DIR "${TPL_ROOT}/conduit-0.6.0" CACHE PATH "")
 
-# MFEM not built
+set(MFEM_DIR "${TPL_ROOT}/mfem-4.1.0" CACHE PATH "")
 
 set(HDF5_DIR "${TPL_ROOT}/hdf5-1.8.21" CACHE PATH "")