+++ src/axom/inlet/tests/inlet_test_utils.cpp
@@ -1,0 +1,10 @@
+// Copyright (c) 2017-2020, Lawrence Livermore National Security, LLC and
+// other Axom Project Developers. See the top-level COPYRIGHT file for details.
+//
+// SPDX-License-Identifier: (BSD-3-Clause)
+
+#include "axom/inlet/tests/inlet_test_utils.hpp"
+#include "axom/slic.hpp"
+
+namespace axom::inlet::detail
+{ }  // namespace axom::inlet::detail